+++ spec/rubocop/cop/style/stderr_puts_spec.rb
@@ -8,14 +8,14 @@ RSpec.describe RuboCop::Cop::Style::StderrPuts do
   it "registers an offense when using `$stderr.puts('hello')`" do
     expect_offense(<<~RUBY)
       $stderr.puts('hello')
-      ^^^^^^^^^^^^ Use `warn` instead of `$stderr.puts` to allow such output to be disabled.
+      ^^^^^^^^^^^^ Use `Kernel.warn` instead of `$stderr.puts` to allow such output to be disabled.
     RUBY
   end
 
-  it 'autocorrects `$stderr.puts` to `warn`' do
+  it 'autocorrects `$stderr.puts` to `Kernel.warn`' do
     new_source = autocorrect_source("$stderr.puts('hello')")
 
-    expect(new_source).to eq "warn('hello')"
+    expect(new_source).to eq "Kernel.warn('hello')"
   end
 
   it 'registers no offense when using `$stderr.puts` with no arguments' do
@@ -27,14 +27,14 @@ RSpec.describe RuboCop::Cop::Style::StderrPuts do
   it "registers an offense when using `STDERR.puts('hello')`" do
     expect_offense(<<~RUBY)
       STDERR.puts('hello')
-      ^^^^^^^^^^^ Use `warn` instead of `STDERR.puts` to allow such output to be disabled.
+      ^^^^^^^^^^^ Use `Kernel.warn` instead of `STDERR.puts` to allow such output to be disabled.
     RUBY
   end
 
-  it 'autocorrects `STDERR.puts` to `warn`' do
+  it 'autocorrects `STDERR.puts` to `Kernel.warn`' do
     new_source = autocorrect_source("STDERR.puts('hello')")
 
-    expect(new_source).to eq "warn('hello')"
+    expect(new_source).to eq "Kernel.warn('hello')"
   end
 
   it 'registers no offense when using `STDERR.puts` with no arguments' do